Move components out
