+++ packages/client/src/game/Bullets.ts
@@ -1,11 +1,11 @@
-import Bullet from "./components/Bullet";
-import Player from "./components/Player";
+import Bullet from "./Bullet";
+import Player from "./Player";
 import Timer from "./Timer";
 
 export default class Bullets {
   protected bullets: Bullet[] = [];
 
-  constructor(protected readonly timer: Timer) {}
+  constructor(protected readonly timer: Timer) { }
 
   spawn(at: Player, angle: number) {
     // TODO: spawn bullet directly in front of gun for more realistic aiming
