+++ packages/client/src/game/client/ClientChat.ts
@@ -1,5 +1,5 @@
 import Chat from "../Chat";
-import Player from "../components/Player";
+import Player from "../Player";
 
 export default class ClientChat extends Chat {
   id: number = 0;
