+++ packages/client/src/game/client/ClientGame.ts
@@ -3,7 +3,6 @@ import type { Connection } from "@smiley-face-game/api";
 import type { Renderer } from "pixi.js";
 import type TileRegistration from "@smiley-face-game/api/tiles/TileRegistration";
 import Game from "../Game";
-import Chat from "../Chat";
 import ClientPlayers from "./ClientPlayers";
 import ClientWorld from "./ClientWorld";
 import ClientNetwork from "./ClientNetwork";
@@ -12,7 +11,7 @@ import ClientSelector from "./ClientSelector";
 import ClientDisplay from "./ClientDisplay";
 import ClientBlockBar from "./ClientBlockBar";
 import Keyboard from "./Keyboard";
-import type ClientPlayer from "./components/ClientPlayer";
+import type ClientPlayer from "./ClientPlayer";
 import { TileLayer } from "@smiley-face-game/api/types";
 import ClientAim from "./ClientAim";
 import ClientBullets from "./ClientBullets";
