+++ packages/client/src/game/interfaces/PlayerCtor.ts
@@ -1,5 +1,5 @@
-import Player from "../components/Player";
+import Player from "../Player";
 
 export default interface PlayerCtor {
-  new (id: number, username: string, isGuest: boolean): Player;
+  new(id: number, username: string, isGuest: boolean): Player;
 }
